requirements.txt
kafka-python
#ess-streaming-data-types >= 0.9.3
